% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-actors.R
\docType{data}
\name{actors}
\alias{actors}
\title{actors database documentation}
\format{
The actors database is a list that contains the
following 1 datasets: CHATHAM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{CHATHAM: }{A dataset with 203 observations and the following
5 variables: Actor, ID, Beg, Country, City.}
}
}
\usage{
actors
}
\description{
actors database documentation
}
\keyword{datasets}
