% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-finance.R
\docType{data}
\name{finance}
\alias{finance}
\title{finance datacube documentation}
\format{
The finance datacube is a list that contains the
following 1 datasets: WHO_GHED.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{WHO_GHED: }{A dataset with 4244 observations and the following
11 variables: stateID, StateName, year, region, che, che_gdp, dom_che, hk_g_gdp, hk_ext_gdp, ext, ext_che.}
}
}
\source{
\itemize{
\item{WHO_GHED: }{
WHO. \emph{Global Health Expenditure Database}. 2024.
}
}
\usage{
finance
}
\description{
finance datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $WHO_GHED
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |            4244|     0     |       0       |
#> |   StateName   |  character  |            4244|    22     |     0.52      |
#> |     year      |    mdate    |            4244|     0     |       0       |
#> |    region     |  character  |            4244|     0     |       0       |
#> |      che      |  character  |            4244|    90     |     2.12      |
#> |    che_gdp    |  character  |            4244|    90     |     2.12      |
#> |    dom_che    |  character  |            4244|    102    |      2.4      |
#> |   hk_g_gdp    |  character  |            4244|   3456    |     81.43     |
#> |  hk_ext_gdp   |  character  |            4244|   3492    |     82.28     |
#> |      ext      |  character  |            4244|    615    |     14.49     |
#> |    ext_che    |  character  |            4244|    615    |     14.49     |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{WHO_GHED: }{
\url{https://apps.who.int/nha/database}
}
}
}

\section{Mapping}{

\itemize{
\item{WHO_GHED: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   country \tab StateName \cr
   code \tab stateID \cr
   } \tab  \cr
   } \tab  \cr
}
}

\keyword{datasets}
