% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships database documentation}
\format{
The memberships database is a list that contains the
following 1 datasets: HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{HUGGO_MEM: }{A dataset with 43442 observations and the following
13 variables: manyID, stateID, Title, Beg, Signature, Force, End,
stateSignature, stateRat, stateForce, stateEnd, StateName, treatyID.}
}
}
\usage{
memberships
}
\description{
memberships database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           43442|     0     |       0       |
#> |    stateID    |  character  |           43442|    20     |     0.05      |
#> |     Title     |  character  |           43442|     0     |       0       |
#> |      Beg      |    mdate    |           43442|    189    |     0.44      |
#> |   Signature   |    mdate    |           43442|    417    |     0.96      |
#> |     Force     |    mdate    |           43442|   33394   |     76.87     |
#> |      End      |    mdate    |           43442|   43195   |     99.43     |
#> |stateSignature |    mdate    |           43442|   23379   |     53.82     |
#> |   stateRat    |    mdate    |           43442|   33928   |     78.1      |
#> |  stateForce   |    mdate    |           43442|   33688   |     77.55     |
#> |   stateEnd    |    mdate    |           43442|   43421   |     99.95     |
#> |   StateName   |  character  |           43442|     2     |       0       |
#> |   treatyID    |  character  |           43442|     2     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
