% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-references.R
\docType{data}
\name{references}
\alias{references}
\title{references datacube documentation}
\format{
The references datacube is a list that contains the
following 2 datasets: GHHR_REF, WHO_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GHHR_REF: }{A dataset with 224 observations and the following
3 variables: Treaty1, Treaty2, RefType.}
\item{WHO_REF: }{A dataset with 65 observations and the following
3 variables: Treaty1, Treaty2, RefType.}
}
}
\source{
\itemize{
\item{GHHR_REF: }{
\link{1} B. M. Meier, O. A. Cabrera, A. Ayala, et al. “Bridging international law and rights-basedlitigation: mapping health-related rights through the development of the global health and human rightsdatabase”. In: \emph{Health & Hum. Rts.} 14 (2012), p. 20.}
\item{WHO_REF: }{
Bhugra D, Pathare S, Joshi R, Kalra G, Torales J, Ventriglio A (2018). “A review of mental healthpolicies from Commonwealth countries.” \emph{International Journal of Social Psychiatry}, \emph{64}(1), 3-8.}
}
}
\usage{
references
}
\description{
references datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GHHR_REF
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    Treaty1    |  character  |             224|     0     |       0       |
#> |    Treaty2    |  character  |             224|     0     |       0       |
#> |    RefType    |  character  |             224|    88     |     39.29     |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $WHO_REF
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    Treaty1    |  character  |              65|     0     |       0       |
#> |    Treaty2    |  character  |              65|     0     |       0       |
#> |    RefType    |  character  |              65|    20     |     30.77     |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{GHHR_REF: }{ \url https://www.globalhealthrights.org/instruments/instrument-region/}
\item{WHO_REF: }{ \url https://www.mindbank.info/collection/un_who_resolutions/all?page=all}
}
}

\keyword{datasets}
