% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube documentation}
\format{
The agreements datacube is a list that contains the
following 3 datasets: WHO, GHHR, HUGGO.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{WHO: }{A dataset with 115 observations and the following
6 variables: manyID, treatyID, Title, Begin, Organisation, Topic.}
\item{GHHR: }{A dataset with 149 observations and the following
8 variables: manyID, treatyID, Title, Begin, Region, LegalStatus, Lineage,
ghhrID.}
\item{HUGGO: }{A dataset with 229 observations and the following
15 variables: manyID, treatyID, Title, Begin, Signature, Force, End, Formal,
TreatyText, url, Dataset, Source, Health_as_primary_intent, Comments, Topic.}
}
}
\source{
\itemize{
\item{WHO: }{
D. Bhugra, S. Pathare, R. Joshi, et al.
“A review of mental health policies from Commonwealthcountries”.
\emph{International Journal of Social Psychiatry} 64.1 (2018), pp. 3-8.}
\item{GHHR: }{
B. M. Meier, O. A. Cabrera, A. Ayala, et al.
“Bridging international law and rights-based litigation:mapping health-related rights through the development of the global health and human rights database”.
\emph{Health & Human Rights.} 14 (2012), p. 20.}
\item{HUGGO: }{
J. Hollway. Health agreements for manydata. 2021.}
}
}
\usage{
agreements
}
\description{
agreements datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $WHO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             114|     0     |       0       |
#> |   treatyID    |  character  |             114|     0     |       0       |
#> |     Title     |  character  |             114|     0     |       0       |
#> |     Begin     |    mdate    |             114|     0     |       0       |
#> | Organisation  |  character  |             114|     0     |       0       |
#> |     Topic     |  character  |             114|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHHR
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             149|     0     |       0       |
#> |   treatyID    |  character  |             149|     0     |       0       |
#> |     Title     |  character  |             149|     0     |       0       |
#> |     Begin     |    mdate    |             149|     3     |     2.01      |
#> |    Region     |  character  |             149|     0     |       0       |
#> |  LegalStatus  |  character  |             149|     0     |       0       |
#> |    Lineage    |  character  |             149|    52     |     34.9      |
#> |    ghhrID     |  character  |             149|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> -------------------------------------------------------------------------------------
#> |      Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -------------------------------------------------------------------------------------
#> |         manyID         |  character  |             229|     0     |       0       |
#> |        treatyID        |  character  |             229|     0     |       0       |
#> |         Title          |  character  |             229|     0     |       0       |
#> |         Begin          |    mdate    |             229|     0     |       0       |
#> |       Signature        |    mdate    |             229|     0     |       0       |
#> |         Force          |    mdate    |             229|    169    |     73.8      |
#> |          End           |    mdate    |             229|     0     |       0       |
#> |         Formal         |  character  |             229|     0     |       0       |
#> |       TreatyText       |  character  |             229|     0     |       0       |
#> |          url           |  character  |             229|    104    |     45.41     |
#> |        Dataset         |  character  |             229|     0     |       0       |
#> |         Source         |  character  |             229|     1     |     0.44      |
#> |Health_as_primary_intent|  character  |             229|     0     |       0       |
#> |        Comments        |  character  |             229|    206    |     89.96     |
#> |         Topic          |  character  |             229|     0     |       0       |
#> -------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{WHO: }{
\url{https://www.mindbank.info/collection/un_who_resolutions/all?page=all}
}
\item{GHHR: }{
\url{https://www.globalhealthrights.org/instruments/instrument-region/}
}
\item{HUGGO: }{Hand-coded data by the GGO team}
}
}

\section{Mapping}{

\itemize{
\item{WHO: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
\item{GHHR: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Year of adoption \tab Begin \cr
}


}
\item{HUGGO: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
