% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements datacube documentation}
\format{
The agreements datacube is a list that contains the
following 3 datasets: WHO, GHHR, HUGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{WHO: }{A dataset with 108 observations and the following
7 variables: manyID, treatyID, Title, Begin, Organisation, Topic, Lineage.}
\item{GHHR: }{A dataset with 149 observations and the following
7 variables: manyID, treatyID, Title, Begin, Region, LegalStatus, Lineage.}
\item{HUGGO: }{A dataset with 254 observations and the following
18 variables: manyID, treatyID, Title, Begin, Signature, Force, End,
Organisation, Topic, Region, LegalStatus, Lineage, Memb.conditions,
Memb.procedures, Text, Comments, url, Source.}
}
}
\source{
\itemize{
\item{WHO: }{
\link{1} D. Bhugra, S. Pathare, R. Joshi, et al. “A review of mental health policies from Commonwealth countries”. In: \emph{International Journal of Social Psychiatry} 64.1 (2018), pp. 3-8.}
\item{GHHR: }{
\link{1} B. M. Meier, O. A. Cabrera, A. Ayala, et al. “Bridging international law and rights-based litigation: mapping health-related rights through the development of the global health and human rights database”. In: \emph{Health & Hum. Rts.} 14 (2012), p. 20.}
\item{HUGGO: }{
\link{1} J. Hollway. Health agreements for manydata. (2022).}
}
}
\usage{
agreements
}
\description{
agreements datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $WHO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             108|     0     |       0       |
#> |   treatyID    |  character  |             108|     0     |       0       |
#> |     Title     |  character  |             108|     0     |       0       |
#> |     Begin     |    mdate    |             108|     0     |       0       |
#> | Organisation  |  character  |             108|     0     |       0       |
#> |     Topic     |  character  |             108|     0     |       0       |
#> |    Lineage    |  character  |             108|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHHR
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             149|     0     |       0       |
#> |   treatyID    |  character  |             149|     0     |       0       |
#> |     Title     |  character  |             149|     0     |       0       |
#> |     Begin     |    mdate    |             149|     0     |       0       |
#> |    Region     |  character  |             149|     0     |       0       |
#> |  LegalStatus  |  character  |             149|     0     |       0       |
#> |    Lineage    |  character  |             149|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             254|     0     |       0       |
#> |   treatyID    |  character  |             254|     0     |       0       |
#> |     Title     |  character  |             254|     0     |       0       |
#> |     Begin     |    mdate    |             254|     0     |       0       |
#> |   Signature   |    mdate    |             254|     0     |       0       |
#> |     Force     |    mdate    |             254|    185    |     72.83     |
#> |      End      |    mdate    |             254|    251    |     98.82     |
#> | Organisation  |  character  |             254|    148    |     58.27     |
#> |     Topic     |  character  |             254|    148    |     58.27     |
#> |    Region     |  character  |             254|    104    |     40.94     |
#> |  LegalStatus  |  character  |             254|    104    |     40.94     |
#> |    Lineage    |  character  |             254|    82     |     32.28     |
#> |Memb.conditions|  character  |             254|    74     |     29.13     |
#> |Memb.procedures|  character  |             254|    210    |     82.68     |
#> |     Text      |  character  |             254|     0     |       0       |
#> |   Comments    |  character  |             254|    231    |     90.94     |
#> |      url      |  character  |             254|     7     |     2.76      |
#> |    Source     |  character  |             254|     1     |     0.39      |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{WHO: }{ \url https://www.mindbank.info/collection/un_who_resolutions/all?page=all}
\item{GHHR: }{ \url https://www.globalhealthrights.org/instruments/instrument-region/}
\item{HUGGO: }{ \url "Hand-coded by the GGO team")}
}
}

\keyword{datasets}
