% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 3 datasets: WHO, GHHR, HUGGO.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{WHO: }{A dataset with 109 observations and the following
7 variables: manyID, Title, Begin, Organisation, Topic, Lineage, treatyID.}
\item{GHHR: }{A dataset with 151 observations and the following
7 variables: manyID, Title, Begin, Region, LegalStatus, Lineage, treatyID.}
\item{HUGGO: }{A dataset with 255 observations and the following
18 variables: manyID, Title, Begin, Signature, Force, End, Organisation,
Topic, Region, LegalStatus, Lineage, Memb.conditions, Memb.procedures,
treatyID, Text, Comments, url, Source.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $WHO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             108|     0     |       0       |
#> |   treatyID    |  character  |             108|     0     |       0       |
#> |     Title     |  character  |             108|     0     |       0       |
#> |     Begin     |    mdate    |             108|     0     |       0       |
#> | Organisation  |  character  |             108|     0     |       0       |
#> |     Topic     |  character  |             108|     0     |       0       |
#> |    Lineage    |  character  |             108|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHHR
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             149|     0     |       0       |
#> |   treatyID    |  character  |             149|     0     |       0       |
#> |     Title     |  character  |             149|     0     |       0       |
#> |     Begin     |    mdate    |             149|     0     |       0       |
#> |    Region     |  character  |             149|     0     |       0       |
#> |  LegalStatus  |  character  |             149|     0     |       0       |
#> |    Lineage    |  character  |             149|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $HUGGO
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |             254|     0     |       0       |
#> |   treatyID    |  character  |             254|     0     |       0       |
#> |     Title     |  character  |             254|     0     |       0       |
#> |     Begin     |    mdate    |             254|     0     |       0       |
#> |   Signature   |    mdate    |             254|     0     |       0       |
#> |     Force     |    mdate    |             254|    185    |     72.83     |
#> |      End      |    mdate    |             254|    251    |     98.82     |
#> | Organisation  |  character  |             254|    148    |     58.27     |
#> |     Topic     |  character  |             254|    148    |     58.27     |
#> |    Region     |  character  |             254|    104    |     40.94     |
#> |  LegalStatus  |  character  |             254|    104    |     40.94     |
#> |    Lineage    |  character  |             254|    82     |     32.28     |
#> |Memb.conditions|  character  |             254|    74     |     29.13     |
#> |Memb.procedures|  character  |             254|    210    |     82.68     |
#> |     Text      |  character  |             254|     0     |       0       |
#> |   Comments    |  character  |             254|    231    |     90.94     |
#> |      url      |  character  |             254|     7     |     2.76      |
#> |    Source     |  character  |             254|     1     |     0.39      |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
