% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 2 datasets: WHO, GHHR.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{WHO: }{A dataset with 109 observations and the following
7 variables: manyID, Title, Beg, Organisation, Topic, Lineage, treatyID.}
\item{GHHR: }{A dataset with 151 observations and the following
7 variables: manyID, Title, Beg, Region, LegalStatus, Lineage, treatyID.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
