% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-agreements.R
\docType{data}
\name{agreements}
\alias{agreements}
\title{agreements database documentation}
\format{
The agreements database is a list that contains the
following 1 datasets: HEIDI.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{HEIDI: }{A dataset with 2280 observations and the following
4 variables: ID, Title, Beg, Signature.}
}
}
\usage{
agreements
}
\description{
agreements database documentation
}
\keyword{datasets}
