% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 1 datasets: HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{HUGGO_MEM: }{A dataset with 39989 observations and the following
13 variables: manyID, stateID, Title, Begin, Signature, Force, End,
stateSignature, stateRat, stateForce, stateEnd, treatyID, StateName.}
}
}
\source{
\itemize{
\item{HUGGO_MEM: }{
\link{1} J. Hollway. Health agreements for manydata. (2022).}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           39989|     0     |       0       |
#> |    stateID    |  character  |           39989|    21     |     0.05      |
#> |     Title     |  character  |           39989|     0     |       0       |
#> |     Begin     |    mdate    |           39989|     0     |       0       |
#> |   Signature   |    mdate    |           39989|     0     |       0       |
#> |     Force     |    mdate    |           39989|   32607   |     81.54     |
#> |      End      |    mdate    |           39989|   39742   |     99.38     |
#> |stateSignature |    mdate    |           39989|   6754    |     16.89     |
#> |   stateRat    |    mdate    |           39989|   34121   |     85.33     |
#> |  stateForce   |    mdate    |           39989|   33887   |     84.74     |
#> |   stateEnd    |    mdate    |           39989|   39736   |     99.37     |
#> |   treatyID    |  character  |           39989|     0     |       0       |
#> |   StateName   |  character  |           39989|     1     |       0       |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{HUGGO_MEM: }{ \url Hand-coded by the GGO team}
}
}

\keyword{datasets}
