% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 1 datasets: HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{HUGGO_MEM: }{A dataset with 35575 observations and the following
17 variables: manyID, Title, Begin, stateID, StateName, StateSignature,
StateRatification, StateForce, StateEnd, Rat=Notif, Accession, Succession,
treatyID, Signature, Force, End, Coder.}
}
}
\source{
\itemize{
\item{HUGGO_MEM: }{
J. Hollway. Health agreements for manydata. 2021.}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO_MEM
#> ------------------------------------------------------------------------------
#> |   Column Name   |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ------------------------------------------------------------------------------
#> |     manyID      |  character  |           35575|     0     |       0       |
#> |      Title      |  character  |           35575|     0     |       0       |
#> |      Begin      |    mdate    |           35575|     0     |       0       |
#> |     stateID     |  character  |           35575|     0     |       0       |
#> |    StateName    |  character  |           35575|     0     |       0       |
#> | StateSignature  |    mdate    |           35575|   5161    |     14.51     |
#> |StateRatification|    mdate    |           35575|   31325   |     88.05     |
#> |   StateForce    |    mdate    |           35575|   31072   |     87.34     |
#> |    StateEnd     |    mdate    |           35575|     0     |       0       |
#> |    Rat=Notif    |  character  |           35575|   34430   |     96.78     |
#> |    Accession    |    mdate    |           35575|   29755   |     83.64     |
#> |   Succession    |  character  |           35575|     0     |       0       |
#> |    treatyID     |  character  |           35575|     0     |       0       |
#> |    Signature    |    mdate    |           35575|     0     |       0       |
#> |      Force      |    mdate    |           35575|   30672   |     86.22     |
#> |       End       |    mdate    |           35575|     0     |       0       |
#> |      Coder      |  character  |           35575|     0     |       0       |
#> ------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{HUGGO_MEM: }{Hand-coded data by the GGO team}
}
}

\section{Mapping}{

\itemize{
\item{HUGGO_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
