% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-memberships.R
\docType{data}
\name{memberships}
\alias{memberships}
\title{memberships datacube documentation}
\format{
The memberships datacube is a list that contains the
following 1 datasets: HUGGO_MEM.
For more information and references to each of the datasets used,
please use the \code{manydata::call_sources()} and \code{manydata::compare_dimensions()} functions.
\describe{
\item{HUGGO_MEM: }{A dataset with 39392 observations and the following
15 variables: manyID, stateID, Title, Begin, Signature, Force, End,
stateSignature, stateRat, stateForce, stateEnd, treatyID, StateName,
Succession, Formal.}
}
}
\source{
\itemize{
\item{HUGGO_MEM: }{
J. Hollway. Health agreements for manydata. 2021.}
}
}
\usage{
memberships
}
\description{
memberships datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $HUGGO_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |    manyID     |  character  |           39392|     0     |       0       |
#> |    stateID    |  character  |           39392|     0     |       0       |
#> |     Title     |  character  |           39392|     0     |       0       |
#> |     Begin     |    mdate    |           39392|     0     |       0       |
#> |   Signature   |    mdate    |           39392|     0     |       0       |
#> |     Force     |    mdate    |           39392|   33054   |     83.91     |
#> |      End      |    mdate    |           39392|   39145   |     99.37     |
#> |stateSignature |    mdate    |           39392|   6683    |     16.97     |
#> |   stateRat    |    mdate    |           39392|   33574   |     85.23     |
#> |  stateForce   |    mdate    |           39392|   33335   |     84.62     |
#> |   stateEnd    |    mdate    |           39392|   39144   |     99.37     |
#> |   treatyID    |  character  |           39392|     1     |       0       |
#> |   StateName   |  character  |           39392|     0     |       0       |
#> |  Succession   |  character  |           39392|     0     |       0       |
#> |    Formal     |  character  |           39392|   16969   |     43.08     |
#> ----------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{HUGGO_MEM: }{Hand-coded data by the GGO team}
}
}

\section{Mapping}{

\itemize{
\item{HUGGO_MEM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   NA \tab NA \cr
}


}
}
}

\keyword{datasets}
