% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations database documentation}
\format{
The organizations database is a list that contains the
following 2 datasets: CHATHAM, IHEID.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{CHATHAM: }{A dataset with 203 observations and the following
5 variables: Actor, OrganizationID, Beg, CountryID, City.}
\item{IHEID: }{A dataset with 124 observations and the following
5 variables: Abbreviation, Areas, organizationID, Beg, Place.}
}
}
\usage{
organizations
}
\description{
organizations database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
