% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations database documentation}
\format{
The organizations database is a list that contains the
following 3 datasets: CHATHAM, IHEID, GHS.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
}
\usage{
organizations
}
\description{
organizations database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $CHATHAM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |     Actor     |  character  |             203|     0     |       0       |
#> |OrganizationID |  character  |             203|     0     |       0       |
#> |      Beg      |   messydt   |             203|     0     |       0       |
#> |   CountryID   |  character  |             203|    136    |      67       |
#> |     City      |  character  |             203|     0     |       0       |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $IHEID
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> | Abbreviation  |  character  |             124|    16     |     12.9      |
#> |     Areas     |  character  |             124|     3     |     2.42      |
#> |organizationID |  character  |             124|     0     |       0       |
#> |      Beg      |   messydt   |             124|    44     |     35.48     |
#> |     Place     |  character  |             124|     3     |     2.42      |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHS
#> --------------------------------------------------------------------------------------
#> |       Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> --------------------------------------------------------------------------------------
#> |      Abbreviation       |  character  |             203|     0     |       0       |
#> |     organizationID      |  character  |             203|     0     |       0       |
#> |           Beg           |    mdate    |             203|     0     |       0       |
#> |          Place          |  character  |             203|     0     |       0       |
#> |          State          |  character  |             203|     0     |       0       |
#> |           URL           |  character  |             203|     0     |       0       |
#> |          Type           |  character  |             203|     0     |       0       |
#> |Health as primary intent?|  character  |             203|     0     |       0       |
#> --------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\keyword{datasets}
