% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-organizations.R
\docType{data}
\name{organizations}
\alias{organizations}
\title{organizations datacube documentation}
\format{
The organizations datacube is a list that contains the
following 3 datasets: CHATHAM, IHEID, GHS.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{CHATHAM: }{A dataset with 203 observations and the following
7 variables: organizationID, Organization, Begin, City, State, Type,
Health_as_primary_intent.}
\item{IHEID: }{A dataset with 124 observations and the following
5 variables: organizationID, Organization, Begin, Areas, City.}
\item{GHS: }{A dataset with 203 observations and the following
8 variables: organizationID, Organization, Begin, City, State, URL, Type,
Health as primary intent?.}
}
}
\source{
\itemize{
\item{CHATHAM: }{
\link{1} S. J. Hoffman, C. B. Cole, and M. Pearcey. \emph{Mapping global health architecture to inform thefuture}. Chatham House London, 2015.}
\item{IHEID: }{
\link{1} \emph{Directory of Global Health Actors}. 2021.}
\item{GHS: }{
\link{1} S. J. Hoffman and C. B. Cole. “Defining the global health system and systematically mapping itsnetwork of actors”. In: \emph{Global Health} 14.38 (2018). DOI: 10.1186/s12992-018-0340-2.}
}
}
\usage{
organizations
}
\description{
organizations datacube documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $CHATHAM
#> -------------------------------------------------------------------------------------
#> |      Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> -------------------------------------------------------------------------------------
#> |     organizationID     |  character  |             203|     0     |       0       |
#> |      Organization      |  character  |             203|     0     |       0       |
#> |         Begin          |    mdate    |             203|     0     |       0       |
#> |          City          |  character  |             203|     0     |       0       |
#> |         State          |  character  |             203|     0     |       0       |
#> |          Type          |   numeric   |             203|     0     |       0       |
#> |Health_as_primary_intent|  character  |             203|     0     |       0       |
#> -------------------------------------------------------------------------------------
#> 
#> 
#> $IHEID
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> ----------------------------------------------------------------------------
#> |organizationID |  character  |             124|    16     |     12.9      |
#> | Organization  |  character  |             124|     0     |       0       |
#> |     Begin     |    mdate    |             124|    44     |     35.48     |
#> |     Areas     |  character  |             124|     3     |     2.42      |
#> |     City      |  character  |             124|     3     |     2.42      |
#> ----------------------------------------------------------------------------
#> 
#> 
#> $GHS
#> --------------------------------------------------------------------------------------
#> |       Column Name       |  Data Type  |  Observations  |  Missing  |  Missing (\%)  |
#> --------------------------------------------------------------------------------------
#> |     organizationID      |  character  |             203|     0     |       0       |
#> |      Organization       |  character  |             203|     0     |       0       |
#> |          Begin          |    mdate    |             203|     0     |       0       |
#> |          City           |  character  |             203|     0     |       0       |
#> |          State          |  character  |             203|     0     |       0       |
#> |           URL           |  character  |             203|     0     |       0       |
#> |          Type           |  character  |             203|     0     |       0       |
#> |Health as primary intent?|  character  |             203|     0     |       0       |
#> --------------------------------------------------------------------------------------
}\if{html}{\out{</div>}}
}
\section{URL}{

\itemize{
\item{CHATHAM: }{ \url https://www.chathamhouse.org/sites/default/files/field/field_document/20150120GlobalHealthArchitectureHoffmanColePearcey.pdf}
\item{IHEID: }{ \url https://www.graduateinstitute.ch/GHC-directories}
\item{GHS: }{ \url https://doi.org/10.1186/s12992-018-0340-2}
}
}

\section{Mapping}{

\itemize{
\item{CHATHAM: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Code \tab organizationID \cr
   Actors \tab Organization \cr
   HQ_states \tab State \cr
   HQ_city \tab City \cr
   Year_inception \tab Begin \cr
}


}
\item{IHEID: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Abbreviation \tab organizationID \cr
   Name \tab Organization \cr
   Start \tab Begin \cr
   'HQ/Offices' \tab City \cr
}


}
\item{GHS: }{
Variable Mapping\tabular{cc}{
   \emph{from} \tab \emph{to} \cr
   Code \tab organizationID \cr
   Actor \tab Organization \cr
   'Year of Inception' \tab Begin \cr
   'Headquarters Location' \tab City \cr
}


}
}
}

\keyword{datasets}
