% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 2 datasets: GHHR_REF, WHO_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GHHR_REF: }{A dataset with 224 observations and the following
3 variables: Treaty1, Treaty2, RefType.}
\item{WHO_REF: }{A dataset with 65 observations and the following
3 variables: Treaty1, Treaty2, RefType.}
}
}
\usage{
references
}
\description{
references database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
