% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-references.R
\docType{data}
\name{references}
\alias{references}
\title{references database documentation}
\format{
The references database is a list that contains the
following 2 datasets: GHHR_REF, WHO_REF.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{GHHR_REF: }{A dataset with 224 observations and the following
3 variables: Treaty1, Treaty2, RefType.}
\item{WHO_REF: }{A dataset with 65 observations and the following
3 variables: Treaty1, Treaty2, RefType.}
}
}
\usage{
references
}
\description{
references database documentation
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{#> $GHHR_REF
#> -- Data Summary ------------------------
#>                            Values
#> Name                       X[[i]]
#> Number of rows             224   
#> Number of columns          3     
#> _______________________          
#> Column type frequency:           
#>   character                3     
#> ________________________         
#> Group variables            None  
#> 
#> -- Variable type: character ----------------------------------------------------
#>   skim_variable n_missing complete_rate min max empty n_unique whitespace
#> 1 Treaty1               0         1      12  29     0      146          0
#> 2 Treaty2               0         1      12  29     0       40          0
#> 3 RefType              88         0.607   2  11     0        3          0
#> 
#> $WHO_REF
#> -- Data Summary ------------------------
#>                            Values
#> Name                       X[[i]]
#> Number of rows             65    
#> Number of columns          3     
#> _______________________          
#> Column type frequency:           
#>   character                3     
#> ________________________         
#> Group variables            None  
#> 
#> -- Variable type: character ----------------------------------------------------
#>   skim_variable n_missing complete_rate min max empty n_unique whitespace
#> 1 Treaty1               0         1      12  25     0       42          0
#> 2 Treaty2               0         1      12  25     0       21          0
#> 3 RefType              20         0.692   2   5     0        2          0
}\if{html}{\out{</div>}}
}
\keyword{datasets}
