% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: AGR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 193 observations and the following
5 variables: Title, Beg, Text_URL, TreatyText, Source.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\subsection{NA}{

NA
}
}
\keyword{datasets}
