% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyhealth-texts.R
\docType{data}
\name{texts}
\alias{texts}
\title{texts database documentation}
\format{
The texts database is a list that contains the
following 1 datasets: AGR_TXT.
For more information and references to each of the datasets used,
please use the \code{data_source()} and \code{data_contrast()} functions.
\describe{
\item{AGR_TXT: }{A dataset with 193 observations and the following
4 variables: Title, Beg, Text_URL, TreatyText.}
}
}
\usage{
texts
}
\description{
texts database documentation
}
\details{
\preformatted{## $CHATHAM
## -- Data Summary ------------------------
##                            Values
## Name                       X[[i]]
## Number of rows             203   
## Number of columns          5     
## _______________________          
## Column type frequency:           
##   character                5     
## ________________________         
## Group variables            None  
## 
## -- Variable type: character --------------------------------------------------------------------------------------------------------------------------------------------------------
## # A tibble: 5 x 8
##   skim_variable n_missing complete_rate   min   max empty n_unique whitespace
## * <chr>             <int>         <dbl> <int> <int> <int>    <int>      <int>
## 1 Actor                 0             1     4    86     0      203          0
## 2 ID                    0             1     3     3     0      203          0
## 3 Beg                   0             1    10    10     0       81          0
## 4 Country               0             1     2     3     0       16          0
## 5 City                  0             1     4    35     0       74          0
}
}
\keyword{datasets}
